@@ -4,7 +4,7 @@
   "main": "webpack.config.js",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --host 0.0.0.0 --poll=500",
+    "start": "ng serve --proxy-config proxy.conf.json -o ---ssl true",
     "build": "ng build",
     "build:prod": "ng build --prod",
     "test": "ng test",
@@ -12,44 +12,44 @@
     "lint:fix": "ng lint --fix",
     "e2e": "ng e2e",
     "build:dev": "webpack-dev-server --inline --hot --progress --port 8080",
-    "build:ngdev": "ng serve"
+    "build:ngdev": "ng serve -o ---ssl true"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
+    "@angular/animations": "^7.2.15",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/core": "^7.2.15",
+    "@angular/forms": "^7.2.15",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/router": "^7.2.15",
+    "@babel/core": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
     "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.17",
-    "@fortawesome/free-regular-svg-icons": "^5.8.1",
-    "@fortawesome/free-solid-svg-icons": "^5.8.1",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.1",
-    "@types/lodash": "^4.14.123",
+    "@fortawesome/fontawesome-svg-core": "^1.2.18",
+    "@fortawesome/free-regular-svg-icons": "^5.8.2",
+    "@fortawesome/free-solid-svg-icons": "^5.8.2",
+    "@ng-bootstrap/ng-bootstrap": "^4.1.2",
+    "@types/lodash": "^4.14.126",
     "angular-router-loader": "^0.8.5",
     "angular2-template-loader": "^0.6.2",
     "automapper-ts": "^1.9.0",
     "awesome-typescript-loader": "^5.2.1",
     "babel": "^6.23.0",
-    "babel-loader": "^8.0.5",
+    "babel-loader": "^8.0.6",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "bootstrap": "^4.3.1",
-    "clean-webpack-plugin": "^2.0.1",
+    "clean-webpack-plugin": "^2.0.2",
     "core-js": "^2.5.4",
     "css-loader": "^2.1.1",
     "font-awesome": "^4.7.0",
     "fontawesome": "^5.6.3",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "jquery": "^3.4.0",
+    "jquery": "^3.4.1",
     "ngx-filter-pipe": "^2.1.2",
     "ngx-markdown": "^7.1.5",
     "npx": "^10.2.0",
@@ -60,15 +60,16 @@
     "tslib": "^1.9.0",
     "tslint-loader": "^3.5.4",
     "uglifyjs-webpack-plugin": "^2.1.2",
-    "webpack": "^4.30.0",
+    "webpack": "^4.31.0",
     "webpack-filter-warnings-plugin": "^1.2.1",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
+    "@angular-devkit/build-angular": "^0.13.9",
+    "@angular/cli": "^7.3.9",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/language-service": "^7.2.15",
+    "@nguniversal/express-engine": "^7.1.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -81,12 +82,14 @@
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "mini-css-extract-plugin": "^0.5.0",
+    "ngx-cookie": "^4.1.2",
+    "pdf-parse": "^1.1.1",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "^5.11.0",
+    "tslint": "^5.16.0",
     "tslint-config-airbnb": "^5.11.1",
     "typescript": "~3.2.2",
-    "webpack-cli": "^3.3.0",
+    "webpack-cli": "^3.3.2",
     "xml2js": "^0.4.19"
   }
 }
